+++ source/c++/include/entity.hpp
