+++ source/c++/include/mathUtils.hpp
