+++ source/c++/include/rect.hpp
