+++ source/c++/include/trajectory.hpp
