+++ source/c++/include/vectorFuncs.hpp
