+++ source/c++/src/entity.cpp
