+++ source/c++/src/entityUtil.cpp
