+++ source/c++/src/mathUtils.cpp
