+++ source/c++/src/tracker.cpp
